refactor m6anet entry point and module names
