+++ m6anet/scripts/train.py
@@ -5,38 +5,60 @@ import joblib
 import toml
 from argparse import ArgumentParser
 from argparse import ArgumentDefaultsHelpFormatter
-from ..utils.training_utils import train, validate
 from ..utils.builder import build_dataloader, build_loss_function
+from..utils.constants import DEFAULT_MODEL_CONFIG
+from ..utils.training_utils import train, validate
 from ..model.model import MILModel
-  
+
 
 def argparser():
     parser = ArgumentParser(
         formatter_class=ArgumentDefaultsHelpFormatter,
         add_help=False
     )
-    parser.add_argument("--model_config", default=None, required=True)
-    parser.add_argument("--train_config", default=None, required=True)
-    parser.add_argument("--save_dir", default=None, required=True)
-    parser.add_argument("--device", default="cpu")
-    parser.add_argument("--lr", default=4e-4, type=float)
-    parser.add_argument("--seed", default=25, type=int)
-    parser.add_argument("--epochs", default=50, type=int)
-    parser.add_argument("--num_workers", default=25, type=int)
-    parser.add_argument("--save_per_epoch", default=10, type=int)
-    parser.add_argument("--weight_decay", dest="weight_decay", default=0, type=float)
-    parser.add_argument("--num_iterations", default=5, type=int)
+    parser.add_argument("--model_config",
+                        help='path to model config file.',
+                        default=DEFAULT_MODEL_CONFIG)
+    parser.add_argument("--train_config",
+                        help='path to training config file.',
+                        required=True)
+    parser.add_argument("--save_dir",
+                        help='directory to output training results.',
+                        required=True)
+    parser.add_argument("--device",
+                        help='device to perform training with.',
+                        default='cpu', type=str)
+    parser.add_argument("--lr",
+                        help='training learning rate.',
+                        default=4e-4, type=float)
+    parser.add_argument("--seed",
+                        help='random seed for training.',
+                        default=25, type=int)
+    parser.add_argument("--epochs",
+                        help='number of training epochs',
+                        default=50, type=int)
+    parser.add_argument("--n_processes",
+                        help='number of processes to use for training.',
+                        default=25, type=int)
+    parser.add_argument("--save_per_epoch",
+                        help='number of epoch multiple to save training checkpoint.',
+                        default=10, type=int)
+    parser.add_argument("--weight_decay",
+                        help='weight decay argument to regularize training.',
+                        default=0, type=float)
+    parser.add_argument("--num_iterations",
+                        help='number of pass during evaluation step',
+                        default=5, type=int)
     return parser
 
 
-def train_and_save(args):
-    
+def main(args):
+
     seed = args.seed
     np.random.seed(seed)
     torch.manual_seed(seed)
 
     device = args.device
-    num_workers = args.num_workers
     n_epoch = args.epochs
     lr = args.lr
     save_per_epoch = args.save_per_epoch
@@ -62,27 +84,30 @@ def train_and_save(args):
     train_info["train_config"]["number_of_validation_iterations"] = n_iterations
     train_info["train_config"]["seed"] = seed
 
-    with open(os.path.join(save_dir, "train_info.toml"), 'w') as f:
+    with open(os.path.join(save_dir, "train_info.toml"), 'w', encoding='utf-8') \
+            as f:
         toml.dump(train_info, f)
-    
-    model = MILModel(model_config).to(device)
-    train_dl, val_dl, test_dl = build_dataloader(train_config, num_workers)
 
-    optimizer = torch.optim.Adam(model.parameters(), lr=lr, weight_decay=weight_decay)
+    model = MILModel(model_config).to(device)
+    train_dl, val_dl, test_dl = build_dataloader(train_config, args.n_processes)
+
+    optimizer = torch.optim.Adam(model.parameters(), lr=lr,
+                                 weight_decay=weight_decay)
 
     criterion = build_loss_function(train_config['loss_function'])
 
-    train_results, val_results = train(model, train_dl, val_dl, optimizer, n_epoch, device, 
+    train_results, val_results = train(model, train_dl, val_dl, optimizer, n_epoch, device,
                                        criterion, save_dir=save_dir,
-                                       save_per_epoch=save_per_epoch, n_iterations=n_iterations)
+                                       save_per_epoch=save_per_epoch,
+                                       n_iterations=n_iterations)
 
-    joblib.dump(train_results, os.path.join(save_dir, "train_results.joblib"))      
-    joblib.dump(val_results, os.path.join(save_dir, "val_results.joblib"))   
+    joblib.dump(train_results, os.path.join(save_dir, "train_results.joblib"))
+    joblib.dump(val_results, os.path.join(save_dir, "val_results.joblib"))
 
     selection_criteria = ['avg_loss', 'roc_auc', 'pr_auc']
-    
+
     for selection_criterion in selection_criteria:
-        val_loss = [val_results[selection_criterion][i] 
+        val_loss = [val_results[selection_criterion][i]
                     for i in range(0, len(val_results[selection_criterion]), save_per_epoch)]
 
         if selection_criterion in ('avg_loss', 'avg_loss_read'):
@@ -103,12 +128,4 @@ def train_and_save(args):
                                                  roc_auc=test_results["roc_auc"],
                                                  pr_auc=test_results["pr_auc"]))
         print("=====================================")
-        joblib.dump(test_results, os.path.join(save_dir, "test_results_{}.joblib".format(selection_criterion)))   
-
-
-def main():
-    args = argparser().parse_args()
-    train_and_save(args)
-
-if __name__ == '__main__':
-    main()
+        joblib.dump(test_results, os.path.join(save_dir, "test_results_{}.joblib".format(selection_criterion)))
