+++ m6anet/utils/helper.py
@@ -8,7 +8,7 @@ from collections import defaultdict
 
 
 class EventalignFile:
-    
+
     def __init__(self, fn):
         self._fn = fn
         self._open()
@@ -52,10 +52,10 @@ def end_queue(task_queue,n_processes):
     for _ in range(n_processes):
         task_queue.put(None)
     return task_queue
-        
+
 def get_ids(f_index,data_info): #todo
     df_list = []
-      
+
     for condition_name, run_names in data_info.items():
         list_of_set_ids = []
         for run_name in run_names:
@@ -65,23 +65,23 @@ def get_ids(f_index,data_info): #todo
         df_list += [pandas.DataFrame({'ids':list(ids),condition_name:[1]*len(ids)})]
     df_merged = reduce(lambda  left,right: pandas.merge(left,right,on=['ids'], how='outer'), df_list).fillna(0).set_index('ids')
     return sorted(list(df_merged[df_merged.sum(axis=1) >= 2].index)) # At least two conditions.
-    
+
 ## tmp: to remove
 # def get_gene_ids(config_filepath): # arguments are not used.
 #     import os,pandas
 #     from ..diffmod.configurator import Configurator
 #     # config_filepath = '/ploy_ont_workspace/github/experiments/Release_v1_0/config_manuscript/gmm_HEK293T-KO_HEK293T-WT_HEPG2-WT_K562-WT_A549-WT_MCF7-WT_reps_v01.ini'
-#     config = Configurator(config_filepath) 
+#     config = Configurator(config_filepath)
 #     paths = config.get_paths()
 #     info = config.get_info()
-#     criteria = config.get_criteria() 
-#     df_gt_ids = pandas.read_csv('/ploy_ont_workspace/out/Release_v1_0/statCompare/data/mapping_gt_ids.csv')    
+#     criteria = config.get_criteria()
+#     df_gt_ids = pandas.read_csv('/ploy_ont_workspace/out/Release_v1_0/statCompare/data/mapping_gt_ids.csv')
 #     gene_ids = set(df_gt_ids['g_id'].unique())
 #     read_count_sum_min,read_count_sum_max = criteria['read_count_min'],criteria['read_count_max']
 #     df_read_count = {}
 #     for run_name in set(info['run_names']):
 #         read_count_filepath = os.path.join(paths['data_dir'],run_name,'summary','read_count_per_gene.csv')
-#         df_read_count[run_name] = pandas.read_csv(read_count_filepath).set_index('g_id') 
+#         df_read_count[run_name] = pandas.read_csv(read_count_filepath).set_index('g_id')
 
 #     for run_name in set(info['run_names']):
 #         df_read_count[run_name].reset_index(inplace=True)
@@ -95,14 +95,14 @@ def get_ids(f_index,data_info): #todo
 
 class Consumer(multiprocessing.Process):
     """ For parallelisation """
-    
+
     def __init__(self,task_queue,task_function,locks=None,result_queue=None):
         multiprocessing.Process.__init__(self)
         self.task_queue = task_queue
         self.locks = locks
         self.task_function = task_function
         self.result_queue = result_queue
-        
+
     def run(self):
         proc_name = self.name
         while True:
@@ -110,7 +110,7 @@ class Consumer(multiprocessing.Process):
             if next_task_args is None:
                 self.task_queue.task_done()
                 break
-            result = self.task_function(*next_task_args,self.locks)    
+            result = self.task_function(*next_task_args,self.locks)
             self.task_queue.task_done()
             if self.result_queue is not None:
                 self.result_queue.put(result)
@@ -130,4 +130,4 @@ def read_last_line(filepath): # https://stackoverflow.com/questions/3346430/what
 
 def is_successful(filepath):
     return read_last_line(filepath) == b'--- SUCCESSFULLY FINISHED ---\n'
-    
+
